[OBP-96] Fix an issue with validating long values.
SVN: 25431

+++ openbis/source/java/ch/systemsx/cisd/openbis/generic/shared/util/SimplePropertyValidator.java
@@ -217,7 +217,7 @@ public class SimplePropertyValidator
             assert value != null : "Unspecified value.";
             try
             {
-                Integer.parseInt(value);
+                Long.parseLong(value);
                 return value;
             } catch (final NumberFormatException ex)
             {
